requirements.txt
Django==3.1.6
graphene-django==2.15.0
